@@ -6,7 +6,7 @@
   "scripts": {
     "analyze": "ANALYZE=true next build",
     "dev": "next",
-    "build": "next build && node ./scripts/generateRSS.js && node ./scripts/generateRedirects.js",
+    "build": "next build && node ./scripts/generateRSS.js && node ./scripts/generateRedirects.js && node ./scripts/downloadFonts.js",
     "lint": "next lint",
     "lint:fix": "next lint --fix",
     "format:source": "prettier --config .prettierrc --write \"{plugins,src}/**/*.{js,ts,jsx,tsx,css}\"",
@@ -22,7 +22,11 @@
     "check-all": "npm-run-all prettier lint:fix tsc"
   },
   "dependencies": {
+<<<<<<< HEAD
     "@codesandbox/sandpack-react": "v0.19.8-experimental.0",
+=======
+    "@codesandbox/sandpack-react": "v0.19.8-experimental.4",
+>>>>>>> 6d965422a4056bac5f93f92735364cb08bcffc6b
     "@docsearch/css": "3.0.0-alpha.41",
     "@docsearch/react": "3.0.0-alpha.41",
     "@headlessui/react": "^1.3.0",
@@ -33,23 +37,24 @@
     "debounce": "^1.2.1",
     "ga-lite": "^2.1.4",
     "github-slugger": "^1.3.0",
-    "next": "^12.0.10",
+    "next": "12.1.7-canary.11",
     "parse-numeric-range": "^1.2.0",
-    "react": "experimental",
+    "react": "0.0.0-experimental-82c64e1a4-20220520",
     "react-collapsed": "3.1.0",
-    "react-dom": "experimental",
+    "react-dom": "0.0.0-experimental-82c64e1a4-20220520",
     "scroll-into-view-if-needed": "^2.2.25"
   },
   "devDependencies": {
     "@mdx-js/loader": "^1.6.16",
     "@types/body-scroll-lock": "^2.6.1",
     "@types/classnames": "^2.2.10",
+    "@types/debounce": "^1.2.1",
     "@types/github-slugger": "^1.3.0",
     "@types/mdx-js__react": "^1.5.2",
     "@types/node": "^14.6.4",
     "@types/parse-numeric-range": "^0.0.1",
-    "@types/react": "^16.9.46",
-    "@types/react-dom": "^16.9.8",
+    "@types/react": "^18.0.9",
+    "@types/react-dom": "^18.0.5",
     "@typescript-eslint/eslint-plugin": "2.x",
     "@typescript-eslint/parser": "2.x",
     "asyncro": "^3.0.0",
