@@ -1,7 +1,6 @@
 /*
  * Copyright (c) Facebook, Inc. and its affiliates.
  */
-
 import React from 'react';
 // @ts-ignore
 import {flushSync} from 'react-dom';
@@ -13,14 +12,20 @@ import {
   SandpackReactDevTools,
 } from '@codesandbox/sandpack-react';
 import scrollIntoView from 'scroll-into-view-if-needed';
-
 import cn from 'classnames';
 
 import {IconChevron} from 'components/Icon/IconChevron';
 import {NavigationBar} from './NavigationBar';
 import {Preview} from './Preview';
 import {CustomTheme} from './Themes';
+<<<<<<< HEAD
 import {useSandpackLint} from './utils';
+=======
+import {useSandpackLint} from './useSandpackLint';
+
+// Workaround for https://github.com/reactjs/reactjs.org/issues/4686#issuecomment-1137402613.
+const emptyArray: Array<any> = [];
+>>>>>>> 6d965422a4056bac5f93f92735364cb08bcffc6b
 
 export function CustomPreset({
   isSingleFile,
@@ -33,7 +38,11 @@ export function CustomPreset({
   devToolsLoaded: boolean;
   onDevToolsLoad: () => void;
 }) {
+<<<<<<< HEAD
   const {lintErrors, onLint} = useSandpackLint();
+=======
+  const {lintErrors, lintExtensions} = useSandpackLint();
+>>>>>>> 6d965422a4056bac5f93f92735364cb08bcffc6b
   const lineCountRef = React.useRef<{[key: string]: number}>({});
   const containerRef = React.useRef<HTMLDivElement>(null);
   const {sandpack} = useSandpack();
@@ -66,7 +75,12 @@ export function CustomPreset({
               showInlineErrors
               showTabs={false}
               showRunButton={false}
+<<<<<<< HEAD
               extensions={[onLint]}
+=======
+              extensions={lintExtensions}
+              extensionsKeymap={emptyArray}
+>>>>>>> 6d965422a4056bac5f93f92735364cb08bcffc6b
             />
             <Preview
               className="order-last xl:order-2"
